+++ Tools/Deploy-MinimalJob.ps1
@@ -119,6 +119,7 @@ function Deploy-MinimalJob {
                            -SkipRollbackOnError:$SkipRollbackOnError `
                            -ColorType $ColorType `
                            -Verbose:$VerbosePreference
+
         if ($PassThru) {
             $ret
         } else {