+++ Tools/TestHelpers.ps1
@@ -11,9 +11,9 @@ function Test-Deploy( $deploy, $running = $true, $podCount = 1, $rollbackStatus 
         Write-Warning ($deploy | ConvertTo-Json -Depth 5)
     }
     if ($zeroExitCode) {
-        $deploy.ExitCode | Should -Be 0
+        $LASTEXITCODE | Should -Be 0
     } else {
-        $deploy.ExitCode | Should -Not -Be 0
+        $LASTEXITCODE | Should -Not -Be 0
     }
     $deploy.Running | Should -Be $running
     $deploy.ReleaseName | Should -Be 'test'
@@ -50,8 +50,8 @@ function Test-MainPod( $podStatus, $status = 'Running', $reason = $null) {
     Test-Pod $podStatus -Status $status -ContainerStatus $status -Reason $reason 'test-minimal-*' 'minimal'
 }
 
-function Test-Job($jobStatus, $running = $false, $rollbackStatus = "DeployedOk", $status = 'Completed', $reason = $null) {
+function Test-Job($jobStatus, $running = $false, $rollbackStatus = "DeployedOk", $status = 'Completed', $reason = $null, [switch] $ZeroDeployExitCode) {
     $jobStatus.Count | Should -Be 2
-    Test-Deploy $jobStatus[0] -running $running -PodCount 0 -rollbackStatus $rollbackStatus -expectedStatus $ignoreError
+    Test-Deploy $jobStatus[0] -running $running -PodCount 0 -rollbackStatus $rollbackStatus -expectedStatus $ignoreError -zeroExitCode:$ZeroDeployExitCode
     Test-Pod $jobStatus[1] -status $status -containerStatus $status -reason $reason -nameLike 'test-job-*' -containerName 'pre-install-upgrade-job'
 }